+++ features/fixtures/mazerunner/src/main/java/com/bugsnag/android/mazerunner/scenarios/AppNotRespondingDisabledScenario.kt
@@ -1,7 +1,8 @@
 package com.bugsnag.android.mazerunner.scenarios
 
+import android.os.Looper
+import android.os.Handler
 import android.content.Context
-import com.bugsnag.android.Bugsnag
 import com.bugsnag.android.Configuration
 
 /**
@@ -11,12 +12,14 @@ internal class AppNotRespondingDisabledScenario(config: Configuration,
                                   context: Context) : Scenario(config, context) {
     init {
         config.setAutoCaptureSessions(false)
-        config.detectAnrs = false
     }
 
     override fun run() {
         super.run()
-        Thread.sleep(6000)
+        val main = Handler(Looper.getMainLooper())
+        main.postDelayed({
+            Thread.sleep(50000) // FOREVER
+        }, 1) // A moment of delay so there is something to 'tap' onscreen
     }
 
 }