+++ sdk/src/main/java/com/bugsnag/android/Client.java
@@ -4,6 +4,7 @@ import static com.bugsnag.android.ConfigFactory.MF_BUILD_UUID;
 import static com.bugsnag.android.MapUtils.getStringFromMap;
 
 import com.bugsnag.android.NativeInterface.Message;
+import com.bugsnag.android.ndk.NativeBridge;
 
 import android.app.Activity;
 import android.app.ActivityManager;
@@ -77,6 +78,7 @@ public class Client extends Observable implements Observer {
     final SharedPreferences sharedPrefs;
 
     private final OrientationEventListener orientationListener;
+    private AppNotRespondingMonitor anrMonitor;
 
     /**
      * Initialize a Bugsnag client
@@ -240,27 +242,25 @@ public class Client extends Observable implements Observer {
 
         // Flush any on-disk errors
         errorStore.flushOnLaunch();
+
+        if (config.getDetectNdkCrashes() || config.getDetectAnrs()) {
+            NativeInterface.configureClientObservers(this);
+        }
     }
 
     private void enableAnrDetection() {
-        long thresholdMs = config.getAnrThresholdMs();
-        Looper looper = Looper.getMainLooper();
-        BlockedThreadDetector.Delegate delegate = new BlockedThreadDetector.Delegate() {
+        AppNotRespondingMonitor.Delegate delegate = new AppNotRespondingMonitor.Delegate() {
             @Override
-            public void onThreadBlocked(Thread thread) {
-                String errMsg = "Application did not respond for at least "
-                    + config.getAnrThresholdMs() + " ms";
+            public void onAppNotResponding(Thread thread) {
+                String errMsg = "Application did not respond to UI input";
                 BugsnagException exc = new BugsnagException("ANR", errMsg, thread.getStackTrace());
 
                 cacheAndNotify(exc, Severity.ERROR, new MetaData(),
                     HandledState.REASON_ANR, null, thread);
             }
         };
-
-        ForegroundDetector foregroundDetector = new ForegroundDetector(appContext);
-        BlockedThreadDetector detector =
-            new BlockedThreadDetector(thresholdMs, looper, foregroundDetector, delegate);
-        detector.start(); // begin monitoring for ANRs
+        anrMonitor = new AppNotRespondingMonitor(delegate);
+        anrMonitor.start(); // begin monitoring for ANRs
     }
 
     @SuppressWarnings("deprecation")
@@ -284,7 +284,15 @@ public class Client extends Observable implements Observer {
 
     void sendNativeSetupNotification() {
         setChanged();
-        super.notifyObservers(new Message(NativeInterface.MessageType.INSTALL, config));
+        ArrayList<Object> messageArgs = new ArrayList<>();
+        messageArgs.add(config);
+        if (anrMonitor != null) {
+            Object buffer = anrMonitor.getSentinelBuffer();
+            if (buffer != null) {
+                messageArgs.add(buffer);
+            }
+        }
+        super.notifyObservers(new Message(NativeInterface.MessageType.INSTALL, messageArgs));
         try {
             Async.run(new Runnable() {
                 @Override